+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+- Fix PR 56 scalar execution-order bypass ([pm-slack-standup-nafb](https://github.com/unbraind/pm-slack-standup/blob/main/.agents/pm/issues/pm-slack-standup-nafb.toon))
+- The publish-attestation gate misses a publish routed through an unquoted shell variable ([pm-slack-standup-83u3](https://github.com/unbraind/pm-slack-standup/blob/main/.agents/pm/issues/pm-slack-standup-83u3.toon))
+
 ## 2026.8.28 - 2026-08-28
 
 ### Fixed